refactor(profiling): add explicit casts to avoid narrowing conversions
Use explicit static_cast where implicit narrowing conversions occur:
- danger.cc: Cast page calculations to safe_memcpy_return_t
- mirrors.cc: Cast table_size calculation to ssize_t
- greenlets.cc: Cast count to int before adding
- stack_renderer.cpp: Cast microsecond_t to int64_t for push_* calls
- vm.cc: Cast size_t to off_t for ftruncate, riovcnt to int for pwritev
- uploader.cpp: Cast len to std::streamsize for write()
- frame.cc: Cast offsetof result to int
- sampler.cpp: Cast uint64_t to double before subtraction

Fixes clang-tidy bugprone-narrowing-conversions warnings.

+++ ddtrace/internal/datadog/profiling/stack/src/sampler.cpp
@@ -114,7 +114,8 @@ Sampler::adapt_sampling_interval()
                             info.system_time.seconds * 1e6 + info.system_time.microseconds);
 #endif
     auto sampler_thread_delta = static_cast<double>(new_sampler_thread_count - sampler_thread_count);
-    auto process_delta = static_cast<double>(new_process_count - process_count - sampler_thread_delta);
+    auto process_delta =
+      static_cast<double>(new_process_count) - static_cast<double>(process_count) - sampler_thread_delta;
     if (process_delta <= 0) {
         process_delta = 1; // Avoid division by zero or negative values
     }
